signmoid derivative update 2nd try
Upon further review, I noticed that the error was confined to the code for plotting the derivative only, where you use x as input to the sigmoid function with deriv=True. Later in the code, when you calculate the weight updates, you use the sigmoid output itself (L1) as the input to the sigmoid derivative. So instead of editing the sigmoid function, it's probably better to change the function call for plotting the derivative. In other words, my fix should only apply to the plotting part in the beginning of the jupyter notebook, not to the update step, and not to the python script at all.

The improved results with the previous edit were likely only due to the edit making the update more aggressive. 

Sorry for the confusion and thanks for indulging me.